refactor: 폴더구조 업보 청산
