+++ src/basic/service/setTimer.js
@@ -1,0 +1,7 @@
+import { flashSaleItemAlert } from './flashSaleItemAlert';
+import { suggestItemAlert } from './suggestItemAlert';
+
+export const setTimer = () => {
+  flashSaleItemAlert();
+  suggestItemAlert();
+};
