+++ src/basic/store/globalStore.js
