--- src/utils/setTimer.js
@@ -1,7 +1,0 @@
-import { flashSaleItemAlert } from '../utils/flashSaleItemAlert';
-import { suggestItemAlert } from '../utils/suggestItemAlert';
-
-export const setTimer = () => {
-  flashSaleItemAlert();
-  suggestItemAlert();
-};
